requirements.txt
asgiref==3.8.1
Django==5.2.1
sqlparse==0.5.3
tzdata==2025.2
Django>=4.2,<5.3
psycopg2-binary>=2.9.9        
openpyxl                      
pandas                        
django-import-export          
django-widget-tweaks          
